refactor(mobile): separate native chat controller contract
